Implemented all classes

+++ Reesen/src/main/java/com/reesen/Reesen/controller/DriverController.java
@@ -30,17 +30,17 @@ public class DriverController {
     }
 
     @GetMapping(value = "/{page}/{size}")
-    public ResponseEntity<List<Driver>> getDrivers(@PathVariable Integer page, @PathVariable Integer size){
+    public ResponseEntity<List<Driver>> getDrivers(@PathVariable Long page, @PathVariable Long size){
         return new ResponseEntity<List<Driver>>(new ArrayList<>(), HttpStatus.OK);
     }
 
     @GetMapping(value = "/{id}")
-    public ResponseEntity<Driver> getDriver(@PathVariable Integer id){
+    public ResponseEntity<Driver> getDriver(@PathVariable Long id){
         return new ResponseEntity<>(null, HttpStatus.OK);
     }
 
     @PutMapping(value = "/{id}")
-    public ResponseEntity<Driver> updateDriver(@RequestBody Driver driver, @PathVariable Integer id){
+    public ResponseEntity<Driver> updateDriver(@RequestBody Driver driver, @PathVariable Long id){
         Driver updatedDriver = driverService.findOne(id);
         /*
         *  Here we set data
